Added files via upload
First attempt to built a Web Server on the ESP8266 module: it creates a web server on the chip that is accessible by a browser by a given IP address obtained during the connection process to the wifi network. The static web page allows to control a relay to turn on/off the light of the circuit.